+++ uniffi_bindgen/src/bindings/swift/templates/Template-Bridging-Header.h
@@ -11,15 +11,15 @@
 // We ensure they are declared exactly once, with a header guard, UNIFFI_SHARED_H.
 #ifdef UNIFFI_SHARED_H
     // We also try to prevent mixing versions of shared uniffi header structs.
-    // If you add anything to the #else block, you must increment the version in UNIFFI_SHARED_HEADER_V1
-    #ifndef UNIFFI_SHARED_HEADER_V1
+    // If you add anything to the #else block, you must increment the version suffix in UNIFFI_SHARED_HEADER_V2
+    #ifndef UNIFFI_SHARED_HEADER_V2
         #error Combining helper code from multiple versions of uniffi is not supported
-    #endif // ndef UNIFFI_SHARED_HEADER_V1
+    #endif // ndef UNIFFI_SHARED_HEADER_V2
 #else
 #define UNIFFI_SHARED_H
-#define UNIFFI_SHARED_HEADER_V1
+#define UNIFFI_SHARED_HEADER_V2
 // ⚠️ Attention: If you change this #else block (ending in `#endif // def UNIFFI_SHARED_H`) you *must* ⚠️
-// ⚠️ increment the version in all instance of UNIFFI_SHARED_HEADER_V1 in this file.                   ⚠️
+// ⚠️ increment the version suffix in all instances of UNIFFI_SHARED_HEADER_V2 in this file.           ⚠️
 
 typedef struct RustBuffer
 {
@@ -46,7 +46,7 @@ typedef struct RustCallStatus {
 } RustCallStatus;
 
 // ⚠️ Attention: If you change this #else block (ending in `#endif // def UNIFFI_SHARED_H`) you *must* ⚠️
-// ⚠️ increment the version in all instance of UNIFFI_SHARED_HEADER_V1 in this file.                   ⚠️
+// ⚠️ increment the version suffix in all instances of UNIFFI_SHARED_HEADER_V2 in this file.           ⚠️
 #endif // def UNIFFI_SHARED_H
   
 {% for func in ci.iter_ffi_function_definitions() -%}